Added rating flag
Added a custom flag, rating.
Use 0 for off and 1 for on in the following order [General][Mature][Adult].
Eg. 110 for General and Mature posts only, and 001 for Adult posts only.
If there is no posts returned then it is likely that the first 30 submissions returned by IB's API does not have been filtered by the rating flag. You can still subscribe to the rss feed and it will update when a match is found.